analyze_models REFACTOR: models_replay
