+++ src/blitzreplays/replays/analyze_models.py
@@ -30,10 +30,6 @@ from blitzmodels import (
     AccountInfo,
     AccountInfoStats,
     AccountId,
-    EnumVehicleTypeStr,
-    Maps,
-    Map,
-    Tank,
     TankStat,
     TankId,
     WGApi,
@@ -41,16 +37,14 @@ from blitzmodels import (
     WGApiWoTBlitzTankopedia,
     WGApiWoTBlitzTankStats,
 )
-from blitzmodels.wotinspector.wi_apiv2 import Replay, PlayerData
-from blitzmodels.wotinspector.wi_apiv1 import EnumBattleResult
 
 from .args import (
-    EnumGroupFilter,
-    EnumTeamFilter,
+    # EnumGroupFilter,
+    # EnumTeamFilter,
     PlayerFilter,
     StatsType,
-    StatsMeasure,
 )
+from .models_replay import PlayerStats, EnrichedReplay
 
 logger = logging.getLogger()
 error = logger.error
@@ -83,145 +77,6 @@ class PlayerStat(AccountInfo):
         if self.statistics is not None and "all" in self.statistics:
             self.stats = self.statistics["all"]
         return self
-
-
-class PlayerStats(JSONExportable):
-    # stat_type: StatsType
-    account_id: AccountId
-    tank_id: int = 0
-    tier: int = 0
-
-    wr: float = 0
-    avgdmg: float = 0
-    battles: int = 0
-    # tier_wr: float = 0
-    # tier_avgdmg: float = 0
-    # tier_battles: int = 0
-    # tank_wr: float = 0
-    # tank_avgdmg: float = 0
-    # tank_battles: int = 0
-
-    # @property
-    # def key(self) -> str:
-    #     return StatsCache.stat_key(self.account_id, self.tier, self.tank_id)
-
-    @property
-    def stats_type(self) -> StatsType:
-        if self.tank_id == 0 and self.tier == 0:
-            return "player"
-        elif self.tank_id > 0:
-            return "tank"
-        elif self.tier > 0:
-            return "tier"
-        else:
-            raise ValueError("player stats record has invalid value")
-
-    def get(self, stats_type: StatsType, measure: StatsMeasure) -> int | float:
-        """get stats measure from the PlayerStats instance. Returns '0' if error or no stat"""
-        if stats_type != self.stats_type:
-            error(f"wrong type of stats record: {self.stats_type} != {stats_type}")
-        return getattr(self, measure, 0)
-
-    @classmethod
-    def from_tank_stat(cls, ts=TankStat) -> Self:
-        """
-        Create PlayerStats from a TankStat for a tank
-        """
-        # return cls(
-        #     account_id=ts.account_id,
-        #     tank_id=ts.tank_id,
-        #     wr=ts.all.wins / ts.all.battles,
-        #     avgdmg=ts.all.damage_dealt / ts.all.battles,
-        #     battles=ts.all.battles,
-        # )
-        if (pts := cls.from_obj(ts)) is not None:
-            return pts
-        else:
-            raise ValueError(f"could not transform TankStat: {ts.json_src()}")
-
-    @classmethod
-    def from_tank_stats(cls, stats=Iterable[TankStat], tier: int = 0) -> Optional[Self]:
-        """
-        Create an aggregate PlayerStats from a list of Iterable[TankStats]
-
-        Creates tank
-        """
-        try:
-            iter_stats = iter(stats)
-            ts: TankStat = next(iter_stats)
-            # res = cls(
-            #     account_id=ts.account_id,
-            #     tier=tier,
-            #     wr=ts.all.wins / ts.all.battles,
-            #     avgdmg=ts.all.damage_dealt / ts.all.battles,
-            #     battles=ts.all.battles,
-            # )
-            res = cls.from_tank_stat(ts)
-            res.tank_id = 0
-            res.tier = tier
-            for ts in iter_stats:
-                try:
-                    res.battles = res.battles + ts.all.battles
-                    res.wr = res.wr + ts.all.wins
-                    res.avgdmg = res.avgdmg + ts.all.damage_dealt
-                except Exception as err:
-                    error(err)
-            if res.battles > 0:
-                res.wr = res.wr / res.battles
-                res.avgdmg = res.avgdmg / res.battles
-            return res
-        except Exception as err:
-            error(f"{type(err)}: {err}")
-        return None
-
-    # @classmethod
-    # def tank_stats(cls, stats: List[TankStat]) -> List[Self]:
-    #     """Create PlayerStats from WG API Tank Stats for each tank"""
-    #     res: List[Self] = list()
-    #     for ts in stats:
-    #         res.append(cls.from_tank_stat(ts))
-    #     return res
-
-    # @classmethod
-    # def tier_stats(
-    #     cls, account_id: AccountId, stats: List[TankStat], tier: int
-    # ) -> Self:
-    #     """
-    #     Create PlayerStats from WG API Tank Stats
-
-    #     Assumes all the TankStats are for the same tier tanks
-    #     """
-    #     res = cls(account_id=account_id, tier=tier)
-    #     try:
-    #         for ts in stats:
-    #             res.battles = res.battles + ts.all.battles
-    #             res.wr = res.wr + ts.all.wins
-    #             res.avgdmg = res.avgdmg + ts.all.damage_dealt
-
-    #         res.wr = res.wr / res.battles
-    #         res.avgdmg = res.avgdmg / res.battles
-    #     except Exception as err:
-    #         debug(err)
-    #     return res
-
-    # @classmethod
-    # def player_stats(cls, stats: AccountInfo) -> "PlayerStats":
-    #     """ "Create PlayerStats from WG API Tank Stats"""
-
-    #     try:
-    #         if (
-    #             stats.statistics is not None
-    #             and (ps := stats.statistics["all"]) is not None
-    #         ):
-    #             return PlayerStats(
-    #                 account_id=stats.account_id,
-    #                 wr=ps.wins / ps.battles,
-    #                 avgdmg=ps.damage_dealt / ps.battles,
-    #                 battles=ps.battles,
-    #             )
-    #     except KeyError:
-    #         debug(f"no player stats for account_id={stats.account_id}")
-    #     return PlayerStats(account_id=stats.account_id)
 
 
 class PlayerStatsDict(JSONExportableRootDict[AccountId, PlayerStat]):
@@ -638,227 +493,6 @@ class StatsCache:
     #             return self._db[key]
     #         except KeyError:
     #             await sleep(2)
-
-
-class EnrichedPlayerData(PlayerData):
-    tank: Tank | None = None
-    # player stats
-    wr: float = 0
-    avgdmg: float = 0
-    battles: int = 0
-    tier_wr: float = 0
-    tier_avgdmg: float = 0
-    tier_battles: int = 0
-    tank_wr: float = 0
-    tank_avgdmg: float = 0
-    tank_battles: int = 0
-
-    model_config = ConfigDict(
-        extra="allow",
-        populate_by_name=True,
-        validate_assignment=False,
-    )
-
-    def add_stats(self, stats: PlayerStats):
-        match stats.stats_type:
-            case "player":
-                self.wr = stats.wr
-                self.avgdmg = stats.avgdmg
-                self.battles = stats.battles
-            case "tier":
-                self.tier_wr = stats.wr
-                self.tier_avgdmg = stats.avgdmg
-                self.tier_battles = stats.battles
-            case "tank":
-                self.tank_wr = stats.wr
-                self.tank_avgdmg = stats.avgdmg
-                self.tank_battles = stats.battles
-            case other:
-                raise ValueError(f"unknown stats type: {other}")
-
-
-class EnrichedReplay(Replay):
-    """
-    EnrichedReplay to store additional metadata for speed up and simplify replay analysis
-    """
-
-    players_dict: Dict[AccountId, EnrichedPlayerData] = Field(default_factory=dict)
-    player: AccountId = -1
-    plat_mate: List[AccountId] = Field(default_factory=list)
-    battle_tier: int = 0
-    map: Map | None = None
-    top_tier: bool = False
-
-    model_config = ConfigDict(
-        extra="allow",
-        populate_by_name=True,
-        validate_assignment=False,
-    )
-
-    @model_validator(mode="after")
-    def read_players_dict(self) -> Self:
-        for player_data in self.players_data:
-            self.players_dict[player_data.dbid] = EnrichedPlayerData.model_validate(
-                player_data
-            )
-        self.players_data = list()
-        return self
-
-    async def enrich(
-        self,
-        tankopedia: WGApiWoTBlitzTankopedia,
-        maps: Maps,
-        player: AccountId = 0,
-    ):
-        """
-        Prepare the (static) replay data for the particular analysis
-        """
-
-        data: EnrichedPlayerData
-        # add tanks
-        for data in self.players_dict.values():
-            tank_id: TankId = data.vehicle_descr
-            try:
-                tank: Tank = tankopedia[tank_id]
-                data.tank = tank
-                self.battle_tier = max(self.battle_tier, int(tank.tier))
-            except KeyError:
-                debug("could not find tank_id=%d from Tankopedia", tank_id)
-
-        # set player
-        if player <= 0:
-            self.player = self.protagonist
-        else:
-            self.player = player
-
-        if self.player in self.allies:
-            pass
-        elif self.player in self.enemies:  # swap teams
-            tmp_team: list[AccountId] = self.allies
-            self.allies = self.enemies
-            self.enemies = tmp_team
-            if self.battle_result == EnumBattleResult.win:
-                self.battle_result = EnumBattleResult.loss
-            elif self.battle_result == EnumBattleResult.loss:
-                self.battle_result = EnumBattleResult.win
-        else:
-            raise ValueError("no account_id=%d in the replay", self.player)
-
-        # set top_tier
-        if (player_tank := self.players_dict[self.player].tank) is not None:
-            self.top_tier = player_tank.tier == self.battle_tier
-
-        # set platoon mate
-        if self.players_dict[self.player].squad_index is not None:
-            plat_id: int = self.players_dict[self.player].squad_index
-            for player in self.allies:
-                if player == self.player:
-                    continue
-                if self.players_dict[player].squad_index == plat_id:
-                    self.plat_mate = [player]
-
-        # remove player and his plat mate from allies
-        self.allies.remove(self.player)
-        if len(self.plat_mate) == 1:
-            self.allies.remove(self.plat_mate[0])
-
-        try:
-            self.map = maps[self.map_id]
-        except (KeyError, ValueError):
-            error(f"no map (id={self.map_id}) in Maps file")
-
-    def get_players(
-        self,
-        filter: PlayerFilter = PlayerFilter(
-            team=EnumTeamFilter.all, group=EnumGroupFilter.all
-        ),
-    ) -> List[AccountId]:
-        """
-        Get players matching the filter from the replay
-        """
-        players: List[AccountId] = list()
-        try:
-            if filter.team == EnumTeamFilter.player:
-                match filter.group:
-                    case EnumGroupFilter.default:
-                        return [self.player]
-                    case EnumGroupFilter.platoon:
-                        return self.plat_mate
-                    case EnumGroupFilter.all:
-                        return [self.player] + self.plat_mate
-                    case EnumGroupFilter.solo:
-                        if len(self.plat_mate) == 0:
-                            return [self.player]
-
-            elif filter.group == EnumGroupFilter.all:
-                match filter.team:
-                    case EnumTeamFilter.allies:
-                        return [self.player] + self.plat_mate + self.allies
-                    case EnumTeamFilter.enemies:
-                        return self.enemies
-                    case EnumTeamFilter.all:
-                        return (
-                            [self.player] + self.plat_mate + self.allies + self.enemies
-                        )
-
-            elif filter.team == EnumTeamFilter.allies:
-                players = self.allies
-            elif filter.team == EnumTeamFilter.enemies:
-                players = self.enemies
-            elif filter.team == EnumTeamFilter.all:
-                players = self.allies + self.enemies
-
-            res: list[AccountId] = list()
-            data: EnrichedPlayerData
-
-            match filter.group:
-                case EnumGroupFilter.default:
-                    return players
-                case EnumGroupFilter.solo:
-                    return [
-                        player
-                        for player in players
-                        if self.players_dict[player].squad_index is None
-                    ]
-                case EnumGroupFilter.platoon:
-                    return [
-                        player
-                        for player in players
-                        if self.players_dict[player].squad_index is not None
-                    ]
-
-                case (
-                    EnumGroupFilter.tank_destroyer
-                    | EnumGroupFilter.light_tank
-                    | EnumGroupFilter.medium_tank
-                    | EnumVehicleTypeStr.heavy_tank
-                ):
-                    tank_type = EnumVehicleTypeStr[filter.group.name]
-                    for player in players:
-                        data = self.players_dict[player]
-                        if data.tank is not None and data.tank.type == tank_type:
-                            res.append(player)
-                    return res
-
-                case EnumGroupFilter.top:
-                    for player in players:
-                        data = self.players_dict[player]
-                        if data.tank is not None and data.tank.tier == self.battle_tier:
-                            res.append(player)
-                    return res
-                case EnumGroupFilter.bottom:
-                    for player in players:
-                        data = self.players_dict[player]
-                        if data.tank is not None and data.tank.tier < self.battle_tier:
-                            res.append(player)
-                    return res
-        except Exception as err:
-            error(err)
-        return []
-
-    def get_player_data(self, account_id: AccountId) -> EnrichedPlayerData:
-        """Get player_data for account_id. Raise an exception if not found"""
-        return self.players_dict[account_id]
 
     # def get_stats_queries(self, stats_type: StatsType) -> Set[StatsQuery]:
     #     queries: Set[StatsQuery] = set()
